pyproject.toml
[project]
name = "poker"
version = "0.1.0"
description = ""
authors = [
    {name = "Martin Survilo"}
]
package-mode = false
requires-python = ">=3.13"
dependencies = [
    "pokerkit (>=0.6.3,<0.7.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ulid-py (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
